+++ common/transmitbase.hpp
@@ -10,10 +10,10 @@
 typedef std::vector<char> bytevector;
 extern bool transmit_verbose;
 extern volatile bool transmit_throw_on_interrupt;
-extern int transmit_bw_report;
-extern unsigned transmit_stats_report;
+extern unsigned long transmit_bw_report;
+extern unsigned long transmit_stats_report;
 extern std::ostream* transmit_cverb;
-extern size_t transmit_chunk_size;
+extern unsigned long transmit_chunk_size;
 
 static const struct VerboseLogNoEol { VerboseLogNoEol() {} } VerbNoEOL;
